various clean ups and utility additions
