+++ ticdat/__init__.py
@@ -15,6 +15,5 @@ The ticDat library is distributed under the BSD2 open source license.
 """
 
 from ticdat.ticdatfactory import TicDatFactory, freeze_me
-from ticdat.logfile import LogFile
-from ticdat.utils import Sloc
+from ticdat.utils import Sloc, LogFile, Progress
 __all__ = ["TicDatFactory", "freeze_me", "LogFile", "Sloc",]
