+++ ticdat/testing/testsql.py
@@ -2,13 +2,12 @@ import os
 import ticdat.utils as utils
 from ticdat.ticdatfactory import TicDatFactory
 from ticdat.testing.ticdattestutils import dietData, dietSchema, netflowData, netflowSchema, firesException
-from ticdat.testing.ticdattestutils import sillyMeData, sillyMeSchema, makeCleanDir, failToDebugger
-from ticdat.testing.ticdattestutils import makeCleanPath, addNetflowForeignKeys, addDietForeignKeys, flaggedAsRunAlone
+from ticdat.testing.ticdattestutils import sillyMeData, sillyMeSchema, makeCleanDir, fail_to_debugger
+from ticdat.testing.ticdattestutils import makeCleanPath, addNetflowForeignKeys, addDietForeignKeys, flagged_as_run_alone
 import shutil
 import unittest
 
-#uncomment decorator to drop into debugger for assertTrue, assertFalse failures
-#@failToDebugger
+#@fail_to_debugger
 class TestSql(unittest.TestCase):
     canRun = False
     @classmethod
