+++ ticdat/testing/testutils.py
@@ -1,11 +1,12 @@
 import sys
 import unittest
 import ticdat.utils as utils
-from ticdat import LogFile
+from ticdat import LogFile, Progress
 from ticdat.ticdatfactory import TicDatFactory, _ForeignKey, _ForeignKeyMapping
 from ticdat.testing.ticdattestutils import dietData, dietSchema, netflowData, netflowSchema, firesException
-from ticdat.testing.ticdattestutils import sillyMeData, sillyMeSchema, makeCleanDir, failToDebugger, flaggedAsRunAlone
+from ticdat.testing.ticdattestutils import sillyMeData, sillyMeSchema, makeCleanDir, fail_to_debugger, flagged_as_run_alone
 from ticdat.testing.ticdattestutils import assertTicDatTablesSame, DEBUG, addNetflowForeignKeys, addDietForeignKeys
+import os
 import itertools
 import shutil
 
@@ -17,7 +18,7 @@ def _deep_anonymize(x)  :
     return map(_deep_anonymize,x)
 
 #uncomment decorator to drop into debugger for assertTrue, assertFalse failures
-#@failToDebugger
+#@fail_to_debugger
 class TestUtils(unittest.TestCase):
     def _testTdfReproduction(self, tdf):
         def _tdfs_same(tdf, tdf2):
@@ -502,13 +503,22 @@ class TestUtils(unittest.TestCase):
                             {k : map(list, v) for k,v in schema.items()})
 
     def testTen(self):
-        with LogFile("boger.txt") as f:
+        with LogFile(os.path.join(_scratchDir, "boger.txt")) as f:
             f.log_table("boger", [["a", "b", 12]] + [list(range(_))[-3:] for _ in list(range(16))[3:]])
-        with LogFile("boger.txt") as f:
+        with LogFile(os.path.join(_scratchDir, "boger.txt")) as f:
             f.log_table("boger", [["a", "b", 12]] + [list(range(_))[-3:] for _ in list(range(12))[3:]])
-        with LogFile("boger.txt") as f:
+        with LogFile(os.path.join(_scratchDir, "boger.txt")) as f:
             self.assertTrue(self.firesException(lambda : f.log_table("boger",
                     [["a", "b", 12]] + [list(range(_))[-3:] for _ in list(range(12))])))
+
+    def testEleven(self):
+        def do_checks(po):
+            self.assertTrue(po.mip_progress("this", 1, 2))
+            self.assertTrue(self.firesException(lambda : po.mip_progress("this", 2.1, 2)))
+            self.assertTrue(po.numerical_progress("boger", 1))
+            self.assertTrue(self.firesException(lambda : po.numerical_progress("boger", "1")))
+        do_checks(Progress())
+        do_checks(Progress(quiet=True))
 
 _scratchDir = TestUtils.__name__ + "_scratch"
