+++ ticdat/utils.py
@@ -218,3 +218,94 @@ class Sloc(object):
             else:
                 s.sloc = Sloc(s)
             return s
+
+class LogFile(object) :
+    """
+    Utility class for writing log files to the Opalytics Cloud Platform.
+    Also enables writing on-the-fly tables into log files.
+    """
+    def __init__(self, path):
+        self._f = open(path, "w") if path else None
+    def write(self, *args, **kwargs):
+        self._f.write(*args, **kwargs) if self._f else None
+    def __enter__(self):
+        return self
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.close()
+    def close(self):
+        self._f.close()if self._f else None
+    def log_table(self, table_name, seq, formatter = lambda _ : "%s"%_,
+                  max_write = 10) :
+        """
+        Writes a table to the log file. Extremely useful functionality for
+        on the fly errors, warnings and diagnostics.
+        :param log_table : the name to be given to the logged table
+        :param seq: An iterable of iterables. The first iterable
+                    lists the field names for the table. The remaining iterables
+                    list the column values for each row. The outer iterable
+                    is thus of length num_rows + 1, while each of the inner
+                    iterables are of length num_cols.
+        :param formatter: a function used to turn column entries into strings
+        :param max_write: the maximum number of table entries to write
+                          to the actual log file. In the Opalytics Cloud Platform,
+                          the log file will link to a scrollable, sortable grid
+                          with all the table entries.
+        :return:
+        """
+        verify(containerish(seq) and all(map(containerish, seq)),
+               "seq needs to be container of containers")
+        verify(len(seq) >= 1, "seq missing initial header row")
+        verify(max(map(len, seq)) == min(map(len, seq)),
+               "each row of seq needs to be the same length as the header row")
+        self.write("Table %s:\n"%table_name)
+        if len(seq[0]) <= 2:
+            ljust = 30
+        elif len(seq[0]) == 3:
+            ljust = 25
+        elif len(seq[0]) == 4:
+            ljust = 20
+        else:
+            ljust = 18
+        if len(seq) - 1 > max_write:
+          self.write("(Showing first %s entries out of %s in total)\n"
+                     %(max_write, len(seq)-1))
+        for row in list(seq)[:max_write+1]:
+            self.write("".join(formatter(_).ljust(ljust) for _ in row) + "\n")
+        self.write("\n")
+
+class Progress(object):
+    """
+    Utility class for indicating progress to the Opalytics Cloud Platform.
+    Also enables writing on-the-fly tables into log files.
+    """
+    def __init__(self, quiet = False):
+        self._quiet = quiet
+    def numerical_progress(self, theme, progress):
+        """
+        indicate generic progress
+        :param theme: string describing the type of progress being advanced
+        :param progress: numerical indicator to the degree of progress advanced
+        :return: False if GUI indicates solve should gracefully finish, True otherwise
+        """
+        verify(stringish(theme), "type_ needs to be string")
+        verify(numericish(progress), "progress needs to be numerical")
+        if not self._quiet:
+             print "%s:%s"%(theme.ljust(40), "{:.5f}".format(progress))
+        return True
+    def mip_progress(self, theme, lower_bound, upper_bound):
+        """
+        indicate progress towards solving a MIP via converging upper and lower bounds
+        :param theme: string describing the type of MIP solve underway
+        :param lower_bound: the best current lower bound to the MIP objective
+        :param upper_bound: the best current upper bound to the MIP objective
+        :return: False if GUI indicates solve should gracefully finish, True otherwise
+        """
+        verify(stringish(theme), "type_ needs to be string")
+        verify(all(map(numericish, (lower_bound, upper_bound))),
+               "lower_bound, upper_bound need to be numeric")
+        verify(lower_bound * 0.99999 <= upper_bound,
+               "lower_bound can't be bigger than upper_bound")
+        if not self._quiet:
+             print "%s:%s:%s"%(theme.ljust(30), "{:.5f}".format(lower_bound).ljust(20),
+                               "{:.5f}".format(upper_bound))
+        return True
